SOS module added

+++ src/main/kotlin/models/alerts/alert.kt
@@ -1,12 +1,13 @@
 package com.example.models.alerts
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 
 data class alert(
     val AlertId: Int,
     val Alert_type: String,
     val Alert_description: String,
-    val latitude: Double,
-    val longitude: Double,
+    val latitude: BigDecimal,
+    val longitude: BigDecimal,
     val created_at: LocalDateTime = LocalDateTime.now(),
 )
